Split out find() and findOrCreate() into separate protocols.
