+++ Sources/CoreDataPlus/ManagedObjectPlus.swift
@@ -1,2 +1,2 @@
 /// Shortcut for adding all ManagedObject extensions
-public typealias ManagedObjectPlus = ManagedObjectCountable & ManagedObjectDeletable & ManagedObjectSearchable & ManagedObjectFindOrCreateBy
+public typealias ManagedObjectPlus = ManagedObjectCountable & ManagedObjectDeletable & ManagedObjectSearchable & ManagedObjectFindOrCreateBy & ManagedObjectFind
